Consolidated logic for traceback dialog into chat_handler
