+++ codeaide/ui/chat_window.py
@@ -398,45 +398,3 @@ class ChatWindow(QMainWindow):
         for item in self.chat_contents:
             self.add_to_chat(item["sender"], item["message"])
         self.logger.info(f"Loaded {len(self.chat_contents)} messages from chat log")
-
-    @pyqtSlot(str)
-    def display_traceback_dialog(self, traceback_text):
-        self.logger.info(f"display_traceback_dialog called with: {traceback_text}")
-        msg_box = QMessageBox(self)
-        msg_box.setWindowTitle("Error Detected")
-        msg_box.setText("An error was detected in the running script:")
-        msg_box.setInformativeText(traceback_text)
-        msg_box.setIcon(QMessageBox.Warning)
-
-        # Create custom buttons
-        send_button = msg_box.addButton("Request a fix", QMessageBox.ActionRole)
-        ignore_button = msg_box.addButton("Ignore", QMessageBox.RejectRole)
-
-        # Set a fixed width for the dialog
-        msg_box.setFixedWidth(600)
-
-        # Make the dialog resizable
-        msg_box.setSizeGripEnabled(True)
-
-        # Set a monospace font for the traceback text
-        text_browser = msg_box.findChild(QTextEdit)
-        if text_browser:
-            font = QFont("Courier")
-            font.setStyleHint(QFont.Monospace)
-            font.setFixedPitch(True)
-            font.setPointSize(10)
-            text_browser.setFont(font)
-
-        msg_box.exec_()
-
-        if msg_box.clickedButton() == send_button:
-            self.send_traceback_to_agent(traceback_text)
-
-    def send_traceback_to_agent(self, traceback_text):
-        message = (
-            "The following traceback occurred when running the code you just provided:\n\n"
-            f"```\n{traceback_text}\n```\n\n"
-            "Please provide a solution that avoids this error."
-        )
-        self.input_text.setPlainText(message)
-        self.on_submit()
